+++ src/DataType/CategoryRelationService.php
@@ -10,7 +10,6 @@ declare(strict_types=1);
 namespace OxidEsales\GraphQL\Example\DataType;
 
 use OxidEsales\GraphQL\Base\DataType\IDFilter;
-use OxidEsales\GraphQL\Example\DataType\CategoryFilterFactory;
 use OxidEsales\GraphQL\Example\Exception\CategoryNotFound;
 use OxidEsales\GraphQL\Example\Service\CategoryRepository;
 use OxidEsales\Eshop\Application\Model\Category as CategoryModel;
@@ -21,7 +20,7 @@ use TheCodingMachine\GraphQLite\Types\ID;
 /**
  * @ExtendType(class=Category::class)
  */
-class CategoryRelationService
+final class CategoryRelationService
 {
     /** @var CategoryRepository */
     private $repository;
@@ -34,10 +33,12 @@ class CategoryRelationService
     /**
      * @Field()
      */
-    public function getParent(Category $child): ?Category
+    public function getParent(Category $category): ?Category
     {
         try {
-            return $this->repository->getById((string)$child->getParentId());
+            return $this->repository->getById(
+                (string)$category->getParentId()
+            );
         } catch (CategoryNotFound $e) {
             return null;
         }
@@ -47,11 +48,11 @@ class CategoryRelationService
      * @Field()
      * @return Category[]
      */
-    public function getChildren(Category $parent): array
+    public function getChildren(Category $category): array
     {
         return $this->repository->getByFilter(
-            CategoryFilterFactory::createCategoryFilter(
-                new IDFilter($parent->getId())
+            new CategoryFilter(
+                new IDFilter($category->getId())
             )
         );
     }